requirements.txt
flake8==3.7.8
Pillow==6.0.0
pdoc3==0.7.0
